[ALPS03878343] Sensor: Fix build error
[Detail]
Modify mdelay declare type for fix build error

MTK-Commit-Id: 95872305ec8f77aec5e81e0a5b461a0cbd6a0711

CR-Id: ALPS03878343
Feature: Sensor Hub

+++ drivers/misc/mediatek/sensors-1.0/alsps/alsps.c
@@ -377,7 +377,7 @@ static int als_enable_and_batch(void)
 		pr_debug("als set ODR, fifo latency done\n");
 		/* start polling, if needed */
 		if (cxt->als_ctl.is_report_input_direct == false) {
-			int mdelay = cxt->als_delay_ns;
+			unsigned long long mdelay = cxt->als_delay_ns;
 
 			do_div(mdelay, 1000000);
 			/* defaut max polling delay */
